requirements.txt
requests
beautifulsoup4
playwright
openrouter
pdfplumber
python-dotenv
